+++ README.md
@@ -13,3 +13,5 @@
     openpyxl打开xlsm文件很慢，打开文件和保存操作比较废时间。
 
     目前对程序的速度测试，因为程序都较小，多线程和多进程以及协程，都没有明显的性能提升，反而有所下降。
+
+    动态规划法最稳定，，但是计算时间较长。排列组合方法，在数据量少的情况下运行很快，数据量大运行时间就容易很长。
